requirements.txt
requests
python-dotenv
pandas
vaderSentiment
transformers
torch
matplotlib
seaborn
tqdm
